@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Http} from '@angular/http';
+import { HttpClient ,HttpParams } from '@angular/common/http';
 
 import { Restaurant } from "./restaurant/restaurant.model";
 import { MenuItem } from "../restaurant-detail/menu-item/menu-item.model"
@@ -14,30 +14,26 @@ import { ErrorHandler} from '../app.error-handler';
 @Injectable()
 export class RestaurantsService {
 
-    constructor(private http: Http){}
+  constructor(private http: HttpClient){}
 
     restaurants(search?: string): Observable<Restaurant[]>{
-        return this.http.get(`${MEAT_API}/restaurants`, {params:{q: search}})
-        .map(response => response.json())
-        .catch(ErrorHandler.handleError)
+        let params: HttpParams = undefined
+        if(search){
+          params = new HttpParams().set('q', search)
+        }
+      return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, {params:params})
     }
 
     restaurantById(id: string): Observable<Restaurant>{
-        return this.http.get(`${MEAT_API}/restaurants/${id}`)
-        .map(response => response.json())
-        .catch(ErrorHandler.handleError)
+      return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`)
     }
 
     reviewsOfRestaurant(id: string): Observable<any>{
         return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`)
-        .map(response => response.json())
-        .catch(ErrorHandler.handleError)
     }
 
     menuOfRestaurant(id: string): Observable<MenuItem[]>{
-        return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-        .map(response => response.json())
-        .catch(ErrorHandler.handleError)
+      return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`)
 
     }
 }
